refactoring test utils to self_tests_utils
